+++ src/components/dashboard/DashBoard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import NavigationBar from "../NavigationBar";
+import NavigationBar from "../landing-page/NavigationBar";
 const DashBoard = () => {
     return (
         <div>
