+++ src/components/landing-page/LandingPage.css
@@ -2,5 +2,5 @@ body {
     overflow: hidden;
 }
 .center-card {
-    margin-top: 7vw;
+    margin-top: 25vh;
 }
