+++ src/components/landing-page/NavigationBar.js
